KioFuse allows you to mount remote directories into the root hierarchy of your local system, thereby exposing KIO filesystems to POSIX-compliant applications: Firefox, OpenOffice, GNOME apps, shell utilities and more. Another important advantage of KioFuse is that it allows you to arrange the hierarchy of *all* your files (local or remote) in whichever way makes sense to you (just like symliks).
This is the start of the KDE4 port. It currently doesn't do as much as the KDE3 version, but the code is cleaned up. In order to make it a viable 
filesystem, I have to figure out a way to allow simultaneous file access. Currently, if a directory takes a long time to be accessed (simulated by a 
sleep(20) call in ListJobHelper::receiveEntries), all new requests to kioFuseReadDir are stalled. The TODO file contains more info on this problem.